requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dataclasses-json==0.6.3
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
hyperframe==6.0.1
idna==3.6
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
llama-index==0.9.24
llama_cpp_python==0.2.26
lxml==5.1.0
markdownify==0.11.6
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.2
openai==1.6.1
packaging==23.2
pandas==2.1.4
pillow==10.2.0
portalocker==2.8.2
protobuf==4.25.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.7.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
safetensors==0.4.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wrapt==1.16.0
yarl==1.9.4